refactor: web3 factory
